requirements.txt
firebase-admin>=6.2.0
google-cloud-firestore>=2.11.0 